# Conflicts:
#	ScriptRunner/Services/OpenProjectForRunService.cs

+++ ScriptHandler/Services/RenameProjectService.cs
@@ -11,7 +11,7 @@ namespace ScriptHandler.Services
 			string projectPath,
 			string projectName)
 		{
-
+			LoggerService.Inforamtion(this, $"Renaming project \"{projectName}\"");
 			string newProjectName = GetNewName(projectName);
 			if(string.IsNullOrEmpty(newProjectName)) 
 				return null;
@@ -39,11 +39,11 @@ namespace ScriptHandler.Services
 			string[] filesList = Directory.GetFiles(originalDir);
 			foreach (string file in filesList)
 			{
+				LoggerService.Inforamtion(this, $"Copying file \"{file}\"");
 				CopyFile(
 					file,
 					newDir,
-					newProjectName,
-					projectName);
+					newProjectName);
 			}
 
 			projectPath = Path.Combine(newDir, newProjectName + ".prj");
@@ -73,8 +73,7 @@ namespace ScriptHandler.Services
 		private void CopyFile(
 			string file,
 			string newDir,
-			string newProjectName,
-			string projectName)
+			string newProjectName)
 		{
 			string fileName = Path.GetFileName(file);
 			string newPath = Path.Combine(newDir, fileName);
@@ -85,37 +84,12 @@ namespace ScriptHandler.Services
 			else if (extension.ToLower() == ".gprj")
 				newPath = Path.Combine(newDir, newProjectName + ".gprj");
 
+			if (File.Exists(newPath))
+				return;
+
 			File.Copy(file, newPath);
 
-			//if (extension.ToLower() == ".prj" || extension.ToLower() == ".gprj")
-			//{
-			//	HandleProjectFile(
-			//		newPath,
-			//		newProjectName,
-			//		projectName);
-			//}
+			
 		}
-
-		//private void HandleProjectFile(
-		//	string newPath,
-		//	string newProjectName,
-		//	string projectName)
-		//{
-		//	string fileData = null;
-		//	using (StreamReader sr = new StreamReader(newPath))
-		//	{
-		//		fileData = sr.ReadToEnd();
-		//	}
-
-		//	if (string.IsNullOrEmpty(fileData))
-		//		return;
-
-		//	fileData = fileData.Replace(projectName, newProjectName);
-
-		//	using (StreamWriter sw = new StreamWriter(newPath))
-		//	{
-		//		sw.Write(fileData);
-		//	}
-		//}
 	}
 }
